Make request method quit earlier/faster
This makes broadcasting and Ping/Info/Stats requests much faster. If the
timeout is 5s, it previously waited the whole time for it to complete.
Now it just waits 250ms after every further response, then returns
early. If all responders take an initial time of e.g. 3 seconds to
reply, it'll still work, as the 250ms timeout is only started after the
first response.